Move `UnitTestThread` and `Require()` into separate file
It's called UnitTestThread.cpp/hpp.
